test: uri 최상단 수정 및 /members/board 이동
